Create new_window_launch.py
The proper way of using QtDesigner is to create a new file and import the converted design to that file.
main.py is the 'main' file here, which launches dialog.py at the start of the application which launches dash.py on a click of a button. Hiding dialog.py upon dash.py launch.